+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## to be released
 
+* feat: add DBaaS firewall rules support in `Databases` client
+
 ## 0.18.2
 
 * feat: add `apiOperationShow()` method to fetch a single database operation by ID